+++ test/gulp.js
@@ -41,6 +41,17 @@ describe('tasks', function() {
     logs.slice(-4).should.eql([
       '',
       '--- Server Dev v0.8.0',
+      'Static files from tmp are served with pushState on port 3000',
+      '',
+    ])
+    rump.reconfigure({server: {pushState: false}})
+    logs.length = 0
+    console.log = newLog
+    gulp.start('spec:info')
+    console.log = log
+    logs.slice(-4).should.eql([
+      '',
+      '--- Server Dev v0.8.0',
       'Static files from tmp are served on port 3000',
       '',
     ])